+++ docs/changelog.rst
@@ -14,6 +14,7 @@ Changes in 0.9.X - DEV
 - Fixed unpickled documents replacing the global field's list. #888
 - Fixed storage of microseconds in ComplexDateTimeField and unused separator option. #910
 - Django support was removed and will be available as a separate extension. #958
+- Don't send a "cls" option to ensureIndex (related to https://jira.mongodb.org/browse/SERVER-769)
 
 Changes in 0.9.0
 ================